Add quick sanity check that the directory we output exists
Of course it could still be deleted when we exit, but there's little we
can do about it.
If however it is already deleted when we do look for it we can return
our fallback instead - though it'd be nicer to try a different child or
return the parent's CWD instead.